requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.35.0
